# api-docs.json
{"swagger":"2.0","info":{"description":"Standing Settlement Instruction for performing CRUD operations in MongoDB","version":"1.0.0","title":"SSI Microservice","license":{"name":"Apache License Version 2.0","url":"https://www.apache.org/licenses/LICENSE-2.0\""}},"host":"whf00aoi:30015","basePath":"/","tags":[{"name":"ssi-controller","description":"Operations pertaining to add,delete,update,search for ssi"}],"paths":{"/ssi/add":{"post":{"tags":["ssi-controller"],"summary":"Allows you to add SSI","operationId":"addSSIUsingPOST","consumes":["application/json"],"produces":["application/json"],"parameters":[{"in":"body","name":"ssi","description":"ssi","required":true,"schema":{"$ref":"#/definitions/SSI"}}],"responses":{"200":{"description":"OK","schema":{"$ref":"#/definitions/SSI"}},"201":{"description":"Created"},"401":{"description":"Unauthorized"},"403":{"description":"Forbidden"},"404":{"description":"Not Found"}}}},"/ssi/delete/{id}":{"delete":{"tags":["ssi-controller"],"summary":"Delete an existing SSI","operationId":"deleteSSIUsingDELETE","produces":["application/json"],"parameters":[{"name":"id","in":"path","description":"id","required":true,"type":"string"}],"responses":{"200":{"description":"OK"},"204":{"description":"No Content"},"401":{"description":"Unauthorized"},"403":{"description":"Forbidden"}}}},"/ssi/findAll":{"get":{"tags":["ssi-controller"],"summary":"Find all SSI","operationId":"findAllUsingGET","produces":["application/json"],"responses":{"200":{"description":"Successfully retrieved list","schema":{"$ref":"#/definitions/SSI"}},"401":{"description":"You are not authorized to view the resource"},"403":{"description":"Accessing the resource you were trying to reach is forbidden"},"404":{"description":"The resource you were trying to reach is not found"}}}},"/ssi/findBySsiName":{"post":{"tags":["ssi-controller"],"summary":"Find SSI by name","operationId":"findBySsiNameUsingPOST","consumes":["application/json"],"produces":["application/json"],"parameters":[{"in":"body","name":"ssi","description":"ssi","required":true,"schema":{"$ref":"#/definitions/SSI"}}],"responses":{"200":{"description":"Successfully retrieved list","schema":{"$ref":"#/definitions/SSI"}},"201":{"description":"Created"},"401":{"description":"You are not authorized to view the resource"},"403":{"description":"Accessing the resource you were trying to reach is forbidden"},"404":{"description":"The resource you were trying to reach is not found"}}}},"/ssi/update":{"post":{"tags":["ssi-controller"],"summary":"Update already created SSI","operationId":"updateSSIUsingPOST","consumes":["application/json"],"produces":["application/json"],"parameters":[{"in":"body","name":"ssi","description":"ssi","required":true,"schema":{"$ref":"#/definitions/SSI"}}],"responses":{"200":{"description":"OK","schema":{"$ref":"#/definitions/SSI"}},"201":{"description":"Created"},"401":{"description":"Unauthorized"},"403":{"description":"Forbidden"},"404":{"description":"Not Found"}}}}},"definitions":{"SSI":{"type":"object","properties":{"assetclassCode":{"type":"string","description":"Asset Class Code","allowEmptyValue":false},"bsAcc":{"type":"string","description":"Buyer/Seller Account","allowEmptyValue":false},"bsId":{"type":"integer","format":"int32","description":"Buyer/Seller ID","allowEmptyValue":false},"bsIdType":{"type":"string","description":"Buyer/Seller ID Type","allowEmptyValue":false},"bsNameAdd":{"type":"string","description":"Buyer/Seller Name","allowEmptyValue":false},"currencyCode":{"type":"string","description":"Currency Code","allowEmptyValue":false},"id":{"type":"string","description":"The database generated product ID","allowEmptyValue":false},"posCode":{"type":"string","description":"Place of Settlement Code","allowEmptyValue":false},"rdaAcc":{"type":"string","description":"Rec Del Account","allowEmptyValue":false},"rdaId":{"type":"integer","format":"int32","description":"Rec Del ID","allowEmptyValue":false},"rdaIdType":{"type":"string","description":"Rec Del ID Type","allowEmptyValue":false},"rdaNameAdd":{"type":"string","description":"Rec Del Name","allowEmptyValue":false},"ssiName":{"type":"string","description":"SSI Name","allowEmptyValue":false},"transSubTypeCode":{"type":"string","description":"Transaction Sub Type Code","allowEmptyValue":false}},"title":"SSI"}}}